@@ -1,7 +1,8 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
+import reportWebVitals from './reportWebVitals';
 import { mergeStyles } from '@fluentui/react';
 
 // Application Insights - https://github.com/microsoft/ApplicationInsights-JS/tree/master/extensions/applicationinsights-react-js
@@ -41,11 +42,16 @@ mergeStyles({
   },
 });
 
-
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
-      <ConfigContext.Provider value={configData}>
+    <ConfigContext.Provider value={configData}>
         <App />
       </ConfigContext.Provider>
-      </React.StrictMode>, document.getElementById('root'))
+  </React.StrictMode>
+);
 
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
